+++ PcProject/src/pc/vision/gui/tools/ColourThresholdConfigTool.java
@@ -317,17 +317,17 @@ public class ColourThresholdConfigTool implements GUITool {
 		@Override
 		public void processFrame(PixelInfo[][] pixels, BufferedImage frame,
 				Graphics2D debugGraphics, BufferedImage debugOverlay) {
-			int[] lg = pitchConstants.getLeftGoal();
-			int[] rg = pitchConstants.getRightGoal();
+			float[] lg = pitchConstants.getLeftGoal();
+			float[] rg = pitchConstants.getRightGoal();
 			int left = pitchConstants.getPitchLeft();
 			int right = left + pitchConstants.getPitchWidth();
 			debugGraphics.setColor(Color.WHITE);
-			debugGraphics.drawRect(left-5, lg[0], 4, 4);
-			debugGraphics.drawRect(left-5, lg[1], 4, 4);
-			debugGraphics.drawRect(left-5, lg[2], 4, 4);
-			debugGraphics.drawRect(right, rg[0], 4, 4);
-			debugGraphics.drawRect(right, rg[1], 4, 4);
-			debugGraphics.drawRect(right, rg[2], 4, 4);
+			debugGraphics.drawRect(left-5, (int) lg[0], 4, 4);
+			debugGraphics.drawRect(left-5, (int) lg[1], 4, 4);
+			debugGraphics.drawRect(left-5, (int) lg[2], 4, 4);
+			debugGraphics.drawRect(right, (int) rg[0], 4, 4);
+			debugGraphics.drawRect(right, (int) rg[1], 4, 4);
+			debugGraphics.drawRect(right, (int) rg[2], 4, 4);
 		}
 		
 	}
